Coli Detection System!
It's Finicky but its up! Although there seems to be a few obvouis bugs.
It only seems to colision check top left portion of object, a lil odd.
Could do with HashCodes Coli detection, will have to see.

Reorganized way Struct,Enum,Consts work. Now more layered and globally
accessible. Mod way BodyMechanics works, can react to types of force,
framework for types of colisions soon too (support for sending it
exists, not for decoding what type of hashtable yet)

+++ Node.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-using NodeEnum;
+using Enums.Node;
 using Structs;
 
 
@@ -315,7 +315,7 @@ namespace ColiSys
         public static Node operator +(Node s1,S_XY s2)
         {
             Node toRet = s1.CopySelf(copyTypes.copyBoth);
-            toRet.Set(s1.Ret(Bounds.l) + s2.x, s1.Ret(Bounds.u) + s2.y);
+            toRet.Set(s1.Ret(Bounds.l) + (int)s2.x, s1.Ret(Bounds.u) + (int)s2.y);
             return toRet;
 
         }
